Add TableDeviceChunkMetadata for tabel model
