+++ java/tsfile/src/main/java/org/apache/tsfile/file/metadata/AlignedChunkMetadata.java
@@ -19,30 +19,16 @@
 
 package org.apache.tsfile.file.metadata;
 
-import org.apache.tsfile.enums.TSDataType;
 import org.apache.tsfile.file.metadata.statistics.Statistics;
-import org.apache.tsfile.read.common.TimeRange;
-import org.apache.tsfile.read.controller.IChunkLoader;
 
-import java.io.OutputStream;
 import java.io.Serializable;
 import java.util.List;
-import java.util.Optional;
 
-public class AlignedChunkMetadata implements IChunkMetadata {
-
-  // ChunkMetadata for time column
-  private final IChunkMetadata timeChunkMetadata;
-  // ChunkMetadata for all subSensors in the vector
-  private final List<IChunkMetadata> valueChunkMetadataList;
-
-  /** ChunkLoader of metadata, used to create IChunkReader */
-  private IChunkLoader chunkLoader;
+public class AlignedChunkMetadata extends AbstractAlignedChunkMetadata {
 
   public AlignedChunkMetadata(
       IChunkMetadata timeChunkMetadata, List<IChunkMetadata> valueChunkMetadataList) {
-    this.timeChunkMetadata = timeChunkMetadata;
-    this.valueChunkMetadataList = valueChunkMetadataList;
+    super(timeChunkMetadata, valueChunkMetadataList);
   }
 
   @Override
@@ -50,31 +36,6 @@ public class AlignedChunkMetadata implements IChunkMetadata {
     return valueChunkMetadataList.size() == 1 && valueChunkMetadataList.get(0) != null
         ? valueChunkMetadataList.get(0).getStatistics()
         : timeChunkMetadata.getStatistics();
-  }
-
-  @Override
-  public Statistics<? extends Serializable> getTimeStatistics() {
-    return timeChunkMetadata.getStatistics();
-  }
-
-  @Override
-  public Optional<Statistics<? extends Serializable>> getMeasurementStatistics(
-      int measurementIndex) {
-    IChunkMetadata chunkMetadata = valueChunkMetadataList.get(measurementIndex);
-    return Optional.ofNullable(chunkMetadata == null ? null : chunkMetadata.getStatistics());
-  }
-
-  @Override
-  public boolean hasNullValue(int measurementIndex) {
-    long rowCount = getTimeStatistics().getCount();
-    Optional<Statistics<? extends Serializable>> statistics =
-        getMeasurementStatistics(measurementIndex);
-    return statistics.map(stat -> stat.hasNullValue(rowCount)).orElse(true);
-  }
-
-  @Override
-  public int getMeasurementCount() {
-    return valueChunkMetadataList.size();
   }
 
   @Override
@@ -87,128 +48,5 @@ public class AlignedChunkMetadata implements IChunkMetadata {
       }
     }
     return false;
-  }
-
-  @Override
-  public boolean isModified() {
-    return timeChunkMetadata.isModified();
-  }
-
-  @Override
-  public void setModified(boolean modified) {
-    timeChunkMetadata.setModified(modified);
-    for (IChunkMetadata v : valueChunkMetadataList) {
-      if (v != null) {
-        v.setModified(modified);
-      }
-    }
-  }
-
-  @Override
-  public boolean isSeq() {
-    return timeChunkMetadata.isSeq();
-  }
-
-  @Override
-  public void setSeq(boolean seq) {
-    timeChunkMetadata.setSeq(seq);
-    for (IChunkMetadata v : valueChunkMetadataList) {
-      if (v != null) {
-        v.setSeq(seq);
-      }
-    }
-  }
-
-  @Override
-  public long getVersion() {
-    return timeChunkMetadata.getVersion();
-  }
-
-  @Override
-  public void setVersion(long version) {
-    timeChunkMetadata.setVersion(version);
-    for (IChunkMetadata valueChunkMetadata : valueChunkMetadataList) {
-      if (valueChunkMetadata != null) {
-        valueChunkMetadata.setVersion(version);
-      }
-    }
-  }
-
-  @Override
-  public long getOffsetOfChunkHeader() {
-    return timeChunkMetadata.getOffsetOfChunkHeader();
-  }
-
-  @Override
-  public long getStartTime() {
-    return timeChunkMetadata.getStartTime();
-  }
-
-  @Override
-  public long getEndTime() {
-    return timeChunkMetadata.getEndTime();
-  }
-
-  @Override
-  public IChunkLoader getChunkLoader() {
-    return chunkLoader;
-  }
-
-  @Override
-  public boolean needSetChunkLoader() {
-    return chunkLoader == null;
-  }
-
-  @Override
-  public void setChunkLoader(IChunkLoader chunkLoader) {
-    this.chunkLoader = chunkLoader;
-  }
-
-  @Override
-  public void setClosed(boolean closed) {
-    timeChunkMetadata.setClosed(closed);
-    for (IChunkMetadata chunkMetadata : valueChunkMetadataList) {
-      if (chunkMetadata != null) {
-        chunkMetadata.setClosed(closed);
-      }
-    }
-  }
-
-  @Override
-  public TSDataType getDataType() {
-    return timeChunkMetadata.getDataType();
-  }
-
-  @Override
-  public String getMeasurementUid() {
-    return timeChunkMetadata.getMeasurementUid();
-  }
-
-  @Override
-  public void insertIntoSortedDeletions(TimeRange timeRange) {
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  public List<TimeRange> getDeleteIntervalList() {
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  public int serializeTo(OutputStream outputStream, boolean serializeStatistic) {
-    throw new UnsupportedOperationException("VectorChunkMetadata doesn't support serial method");
-  }
-
-  @Override
-  public byte getMask() {
-    return 0;
-  }
-
-  public IChunkMetadata getTimeChunkMetadata() {
-    return timeChunkMetadata;
-  }
-
-  public List<IChunkMetadata> getValueChunkMetadataList() {
-    return valueChunkMetadataList;
   }
 }
