+++ java/tsfile/src/main/java/org/apache/tsfile/file/metadata/TableDeviceTimeSeriesMetadata.java
@@ -24,9 +24,9 @@ import org.apache.tsfile.file.metadata.statistics.Statistics;
 import java.util.Collections;
 import java.util.List;
 
-public class AlignedTimeSeriesMetadata extends AbstractAlignedTimeSeriesMetadata {
+public class TableDeviceTimeSeriesMetadata extends AbstractAlignedTimeSeriesMetadata {
 
-  public AlignedTimeSeriesMetadata(
+  public TableDeviceTimeSeriesMetadata(
       TimeseriesMetadata timeseriesMetadata, List<TimeseriesMetadata> valueTimeseriesMetadataList) {
     super(timeseriesMetadata, valueTimeseriesMetadataList);
   }
@@ -37,21 +37,7 @@ public class AlignedTimeSeriesMetadata extends AbstractAlignedTimeSeriesMetadata
    */
   @Override
   public Statistics getStatistics() {
-    return valueTimeseriesMetadataList.size() == 1 && valueTimeseriesMetadataList.get(0) != null
-        ? valueTimeseriesMetadataList.get(0).getStatistics()
-        : timeseriesMetadata.getStatistics();
-  }
-
-  @Override
-  public boolean timeAllSelected() {
-    for (int index = 0; index < getMeasurementCount(); index++) {
-      if (!hasNullValue(index)) {
-        // When there is any value page point number that is the same as the time page,
-        // it means that all timestamps in time page will be selected.
-        return true;
-      }
-    }
-    return false;
+    return timeseriesMetadata.getStatistics();
   }
 
   @Override
@@ -60,13 +46,11 @@ public class AlignedTimeSeriesMetadata extends AbstractAlignedTimeSeriesMetadata
       IChunkMetadata timeChunkMetadata,
       List<IChunkMetadata> chunkMetadataList,
       boolean exits) {
-    if (exits) {
-      res.add(new AlignedChunkMetadata(timeChunkMetadata, chunkMetadataList));
-    }
+    res.add(new TableDeviceChunkMetadata(timeChunkMetadata, chunkMetadataList));
   }
 
   @Override
   AbstractAlignedChunkMetadata constructOnlyTimeChunkMetadata(IChunkMetadata timeChunkMetadata) {
-    return new AlignedChunkMetadata(timeChunkMetadata, Collections.emptyList());
+    return new TableDeviceChunkMetadata(timeChunkMetadata, Collections.emptyList());
   }
 }
