+++ java/tsfile/src/test/java/org/apache/tsfile/tableview/TableViewTest.java
@@ -20,7 +20,6 @@
 package org.apache.tsfile.tableview;
 
 import org.apache.tsfile.enums.TSDataType;
-import org.apache.tsfile.exception.read.ReadProcessException;
 import org.apache.tsfile.exception.write.WriteProcessException;
 import org.apache.tsfile.file.metadata.IDeviceID;
 import org.apache.tsfile.file.metadata.IDeviceID.Factory;
@@ -107,39 +106,6 @@ public class TableViewTest {
   @Test
   public void testWriteOneTable() throws Exception {
     testWrite(testTableSchema);
-  }
-
-  public static void main(String[] args) throws IOException, ReadProcessException {
-    File testFile =
-        new File(
-            "C:\\Users\\JT\\Downloads\\sequence-root.test_g_0-1-2714-1729258251084-4-0-0.tsfile");
-    TsFileSequenceReader sequenceReader = new TsFileSequenceReader(testFile.getAbsolutePath());
-    TableQueryExecutor tableQueryExecutor =
-        new TableQueryExecutor(
-            new MetadataQuerierByFileImpl(sequenceReader),
-            new CachedChunkLoaderImpl(sequenceReader),
-            TableQueryOrdering.DEVICE);
-
-    final TsBlockReader reader =
-        tableQueryExecutor.query("table_5", Arrays.asList("s_0"), null, null, null);
-    assertTrue(reader.hasNext());
-    int cnt = 0;
-    while (reader.hasNext()) {
-      final TsBlock result = reader.next();
-      for (int i = 0; i < result.getPositionCount(); i++) {
-        String col = result.getColumn(0).getObject(i).toString();
-        StringBuilder builder = new StringBuilder(col);
-        for (int j = 1; j < result.getValueColumns().length; j++) {
-          if (result.getColumn(j).isNull(i)) {
-            builder.append(",").append(result.getColumn(j).getObject(i).toString());
-          } else {
-            builder.append(",").append("null");
-          }
-        }
-        System.out.println(result.getTimeByIndex(i) + "\t" + builder.toString());
-      }
-      cnt += result.getPositionCount();
-    }
   }
 
   public static void writeTsFile(TableSchema tableSchema, File file)
